Enlarge font in reviews also
For reasons I don't really understand, increasing the question label
font size inside `viewDidLoad` only makes the question appear bigger
during lesson quizzes, and does not make the question appear bigger
during reviews. I think it might be because of the setupContext block,
but I wasn't able to fix it by changing that.

Setting the font size to the correct size in `viewWillLayoutSubviews`
always seems to work, so I'm now doing that.

This patch increases the font size for both custom fonts and the system
font, in both lessons and reviews, both when tapping on the question
label and when not tapping on the question label.