fix crash by initializing date ivar
This ivar might not be initialized by the first time we send it off to
the compare function, and in that case we might get a EXC_BAD_ACCESS
crash. I happened to run across the crash from trying to read the
uninitialized date ivar as an NSDate while trying to debug something
else, and it was suuuuper confusing for a minute. 😬