only set TKMAudioPlaying once, before playing
it turns out this KeyValueObserver was firing allll the time, including
every single time the app ever activated. we don't actually want to set
the state machine into "audio is playing now" just because you left the
app and came back while there was an audio file loaded up and able to be
played.

fixes #130