Allow trailing n-kana completion in Gboard Godan layout
This commit enables a single, trailing 'n' entry to be converted to 'ん' on answer submission, to be like other kana input sources, where it otherwise requires entering 'nn' to be complete.

Note that hardware keyboard input on iOS requires two enter presses to complete 'ん' from 'n' with this behavior being the same for other Japanese keyboards.