Support wrong answer counts greater than one
These revisions make changes to wrong answer handling for the following conditions.

- Incorrect answers beyond the first one for meanings and readings now get recorded and submitted to the Wanikani API.
- If, after submitting an incorrect answer and skipping the subject, the answer counts in the review queue persist instead of being discarded.
- When a subject reappears, where previously a reading or meaning has been entered correctly, a prompt for the correct part will not appear.

When studying with grouped meanings and readings (back-to-back), entering an incorrect meaning, viewing its details, and then skipping the subject, it reappears at the end of the review queue. However, wrong answers for the subject not recorded resulted in SRS progress advancing mistakenly. It could reach the "burned" stage even after multiple incorrect guesses during the review.