remove keyboard inset hack, unglitch iPhone 14 Pro
As seen in screenshots on #587, the UITextField for the answer sometimes seems to move itself down without adjusting the color gradient above it or the lesson view below it. in my testing, this seems to happen when switching between english and japanese keyboards automatically, which sometimes (but not always) triggers the inset hack. After the inset hack has triggered, the heights seem to be wrong.

I'm not sure why this seems to trigger so much more often on iPhone 14 Pro, and not before, but deleting the inset hack seems to resolve this issue, in my testing on a physical iPhone 14 Pro.

fixes #587 (I hope)