match WaniKani keyboard shortcuts in details pane (#597)
after missing a reading, WaniKani shows the correct reading, but hides the meaning and the mnemonics. there is [a keyboard shortcut][1] to show the hidden information: spacebar.

This can (sometimes) conflict with an existing keyboard shortcut to play the audio reading for a vocabulary word. I have tried to preserve that behavior everywhere possible, and only take over the shortcut when there is hidden information to show.

This commit makes Tsurukame match the WaniKani website shortcuts "f" (to show details after an incorrect answer), "spacebar" (to show the hidden full information while viewing details), and "j" (to play the audio recording of a vocabulary word while viewing details).

[1]: https://knowledge.wanikani.com/wanikani/keyboard-shortcuts/