Force text back to center after kana replacement (#616)
I have no idea why this is now breaking sometimes on Mac, in both Catalyst and iPad modes. This fixes the text suddenly becoming left-aligned after every replacement of kana, but it does not fix the text suddenly jumping to the left if you click inside the text field.

I wasn't able to figure out how to force the text back to the center when you click on the text field. My attempts to intercept `becomeFirstResponder` and `textFieldDidBeginEditing` didn't work, but I don't know if it's because that isn't the problem or if it's because I am not good enough at UIKit. 😅