Add ‘Critical Alerts’ and ‘Time Sensitive Notifications’ to Notifications
New feature:

Add ‘Critical Alerts’ and ‘Time Sensitive Notifications’ to Notifications - If you haven’t done your reviews within half an hour, even if your phone is muted or Do Not Disturb is on, you will be alerted, if the reviews are completed within the 30-minute window, then cancel this ‘Time Sensitive Notifications’.

Attempt to add under the iPhone Settings App, under my App section, under the Notifications section, "Always Deliver Immediately" section, and under this section, there will be a ‘Critical Alerts' toggle option and there will be a 'Time Sensitive Notifications' toggle option.

Critical alerts appear on the Lock Screen and play a sound even if a Focus is on or iPhone is muted.