Fix text drawing for context sentence blurring
Closes #718

It seems as though something is different in the way text is drawn in ContextSentenceModelItem and AttributedModeltem text measurement. Starting with iOS 16, UITextView changed its behavior for text layout. Reverting that change (to TextKit 1 layout) causes the Japanese text to draw the same for edge cases, but then the drawing in ContextSentence... is different.

To make it draw the same, grab the attributed string from the NSTextView and draw that instead. (The difference between the two attributed strings seems to be the font, interestingly enough, unless there is another attribute that I did not notice - I just checked via a simple print(), which may have skipped something).

This is not the greatest fix in the world, I think, but it works on iOS 17 (SE 3 Simulator), and that is _something_ at least!

Test cases:
分
一台
二台
生まれ
生まれる
写る
足りない
交通
仮に
命
芸者

All these test cases were broken before and are working now.