Add options to burn/resurrect subjects.
This currently doesn't work because the user's web cookie expires after some time. We'd need to login again using the username and password if it has expired.

Also, if the user clicks burn/resurrect then goes back to a list of subjects, we need to update the modified subject in the list.